+++ nomium/shares-logger/src/storage/clickhouse/mod.rs
@@ -1,6 +1,6 @@
 mod service;
 mod queries;
-mod retry;
+mod retry_config;
 pub use service::{ClickhouseStorage, ClickhouseBlockStorage};
 mod pool_manager;
-pub use pool_manager::ConnectionPool;
+pub use pool_manager::ClickhouseConnectionPool;
